Fixed: Figure export fails under KDE. Fixed: Figure export fails if filename contains dots under all OS
App_PyPlane.py:
Under some KDE's the file_type is returned empty
because of a "DBUS-error". Hence, in such cases, we try to take the
file_type from the extension specified by the user . If no valid
extension is set by the user file_type is set to png.

Furthermore under all OS the case is handled now where the user wants
to have dots in the basename of the file.