Fixed: App crashes when options are not found in the default config file
For this purpose the sectionlist in config_description.py had to be extended in such a way that it does not only hold the section names but also the corresponding prefixes of the section. This enables the program to set default values for options in the ConfigHandler object that have not been found in the default config file.